// app/gql/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  BigDecimal: { input: any; output: any };
  BigInt: { input: any; output: any };
  Bytes: { input: any; output: any };
  /**
   * 8 bytes signed integer
   *
   */
  Int8: { input: any; output: any };
};

export enum Aggregation_Interval {
  Day = "day",
  Hour = "hour",
}

export type Auction = {
  __typename?: "Auction";
  bids: Array<Bid>;
  bidsCount: Scalars["Int"]["output"];
  collateralToken?: Maybe<CollateralToken>;
  collateralTokenId: Scalars["BigInt"]["output"];
  currencyToken: CurrencyToken;
  endTime: Scalars["BigInt"]["output"];
  highestBid?: Maybe<Bid>;
  id: Scalars["Bytes"]["output"];
  liquidation: Liquidation;
  quantity: Scalars["BigInt"]["output"];
  status?: Maybe<AuctionStatus>;
};
// ...
export type Batch = {
  __typename?: "Batch";
  collateralWrapperContext: Scalars["Bytes"]["output"];
  id: Scalars["ID"]["output"];
  owner: Scalars["Bytes"]["output"];
  underlyingCollateralTokenAddress: Scalars["Bytes"]["output"];
  underlyingCollateralTokenIds: Array<Scalars["BigInt"]["output"]>;
};
// ...
export type Bid = {
  __typename?: "Bid";
  amount: Scalars["BigInt"]["output"];
  auction: Auction;
  bidder: Scalars["Bytes"]["output"];
  id: Scalars["Bytes"]["output"];
  isHighest: Scalars["Boolean"]["output"];
  timestamp: Scalars["BigInt"]["output"];
  transactionHash: Scalars["Bytes"]["output"];
};
// ...
export type Bundle = {
  __typename?: "Bundle";
  collateralWrapperContext: Scalars["Bytes"]["output"];
  id: Scalars["ID"]["output"];
  owner: Scalars["Bytes"]["output"];
  underlyingCollateralTokenAddress: Scalars["Bytes"]["output"];
  underlyingCollateralTokenIds: Array<Scalars["BigInt"]["output"]>;
};
// ...
export type CollateralLiquidated = {
  __typename?: "CollateralLiquidated";
  id: Scalars["Bytes"]["output"];
  loan: Loan;
  proceeds: Scalars["BigInt"]["output"];
};
// ...
export type CollateralToken = {
  __typename?: "CollateralToken";
  id: Scalars["Bytes"]["output"];
  name: Scalars["String"]["output"];
  pools: Array<Pool>;
};
// ...
export type CurrencyToken = {
  __typename?: "CurrencyToken";
  id: Scalars["Bytes"]["output"];
  name: Scalars["String"]["output"];
  pools: Array<Pool>;
  symbol: Scalars["String"]["output"];
};
// ...
export type Deposit = {
  __typename?: "Deposit";
  account: Scalars["Bytes"]["output"];
  createdAt: Scalars["BigInt"]["output"];
  depositedAmount: Scalars["BigInt"]["output"];
  id: Scalars["Bytes"]["output"];
  pool: Pool;
  redemptions: Array<Redemption>;
  shares: Scalars["BigInt"]["output"];
  tick: Tick;
  updatedAt: Scalars["BigInt"]["output"];
};
// ...
export type Deposited = {
  __typename?: "Deposited";
  account: Scalars["Bytes"]["output"];
  amount: Scalars["BigInt"]["output"];
  id: Scalars["Bytes"]["output"];
  shares: Scalars["BigInt"]["output"];
  tick: Tick;
};
// ...
export type Liquidation = {
  __typename?: "Liquidation";
  auctions: Array<Auction>;
  collateralToken: CollateralToken;
  currencyToken: CurrencyToken;
  id: Scalars["Bytes"]["output"];
  loan: Loan;
  source: Scalars["Bytes"]["output"];
  sourceImplementation: Scalars["Bytes"]["output"];
};
// ...
export type Loan = {
  __typename?: "Loan";
  batch?: Maybe<Batch>;
  borrower: Scalars["Bytes"]["output"];
  bundle?: Maybe<Bundle>;
  collateralToken: CollateralToken;
  collateralTokenIds: Array<Scalars["BigInt"]["output"]>;
  collateralWrapperContext: Scalars["Bytes"]["output"];
  collateralWrapperToken?: Maybe<Scalars["Bytes"]["output"]>;
  collateralWrapperTokenId?: Maybe<Scalars["BigInt"]["output"]>;
  completion?: Maybe<Scalars["BigInt"]["output"]>;
  delegate?: Maybe<Scalars["Bytes"]["output"]>;
  duration: Scalars["BigInt"]["output"];
  id: Scalars["Bytes"]["output"];
  interests: Array<Scalars["BigInt"]["output"]>;
  loanReceipt: Scalars["Bytes"]["output"];
  maturity: Scalars["BigInt"]["output"];
  pool: Pool;
  principal: Scalars["BigInt"]["output"];
  proceeds?: Maybe<Scalars["BigInt"]["output"]>;
  repayment: Scalars["BigInt"]["output"];
  status: LoanStatus;
  ticks: Array<Scalars["BigInt"]["output"]>;
  timestamp: Scalars["BigInt"]["output"];
  transactionHash: Scalars["Bytes"]["output"];
  useds: Array<Scalars["BigInt"]["output"]>;
};

export type LoanLiquidated = {
  __typename?: "LoanLiquidated";
  id: Scalars["Bytes"]["output"];
  loan: Loan;
};
// ...
export type LoanOriginated = {
  __typename?: "LoanOriginated";
  id: Scalars["Bytes"]["output"];
  loan: Loan;
};
// ...
export type LoanRepaid = {
  __typename?: "LoanRepaid";
  id: Scalars["Bytes"]["output"];
  loan: Loan;
};
// ...
export type Pool = {
  __typename?: "Pool";
  adminFeeBalance: Scalars["BigInt"]["output"];
  adminFeeRate: Scalars["BigInt"]["output"];
  collateralLiquidator: Scalars["Bytes"]["output"];
  collateralToken: CollateralToken;
  collateralWrappers: Array<Scalars["Bytes"]["output"]>;
  currencyToken: CurrencyToken;
  dayData: Array<PoolDayData>;
  delegationRegistry: Scalars["Bytes"]["output"];
  deposits: Array<Deposit>;
  durations: Array<Scalars["BigInt"]["output"]>;
  events: Array<PoolEvent>;
  id: Scalars["Bytes"]["output"];
  implementation: Scalars["Bytes"]["output"];
  implementationVersionMajor: Scalars["String"]["output"];
  loans: Array<Loan>;
  loansActive: Scalars["BigInt"]["output"];
  loansCollateralLiquidated: Scalars["BigInt"]["output"];
  loansLiquidated: Scalars["BigInt"]["output"];
  loansOriginated: Scalars["BigInt"]["output"];
  loansRepaid: Scalars["BigInt"]["output"];
  maxBorrows: Array<Scalars["BigInt"]["output"]>;
  rates: Array<Scalars["BigInt"]["output"]>;
  ticks: Array<Tick>;
  tokenIdMerkleMetadataURI?: Maybe<Scalars["String"]["output"]>;
  tokenIdMerkleRoot?: Maybe<Scalars["Bytes"]["output"]>;
  tokenIdRange?: Maybe<Array<Scalars["BigInt"]["output"]>>;
  tokenIdSet?: Maybe<Array<Scalars["BigInt"]["output"]>>;
  totalValueAvailable: Scalars["BigInt"]["output"];
  totalValueLocked: Scalars["BigInt"]["output"];
  totalValueUsed: Scalars["BigInt"]["output"];
};
// ...
export type PoolDayData = {
  __typename?: "PoolDayData";
  id: Scalars["Bytes"]["output"];
  pool: Pool;
  timestamp: Scalars["BigInt"]["output"];
  totalValueAvailable: Scalars["BigInt"]["output"];
  totalValueLocked: Scalars["BigInt"]["output"];
  totalValueUsed: Scalars["BigInt"]["output"];
};
// ...
export type PoolEvent = {
  __typename?: "PoolEvent";
  account: Scalars["Bytes"]["output"];
  collateralLiquidated?: Maybe<CollateralLiquidated>;
  deposit?: Maybe<Deposit>;
  deposited?: Maybe<Deposited>;
  from: Scalars["Bytes"]["output"];
  id: Scalars["Bytes"]["output"];
  loanLiquidated?: Maybe<LoanLiquidated>;
  loanOriginated?: Maybe<LoanOriginated>;
  loanRepaid?: Maybe<LoanRepaid>;
  pool: Pool;
  redeemed?: Maybe<Redeemed>;
  timestamp: Scalars["BigInt"]["output"];
  tokenCreated?: Maybe<TokenCreated>;
  transactionHash: Scalars["Bytes"]["output"];
  type: PoolEventType;
  withdrawn?: Maybe<Withdrawn>;
};
// ...
export type Query = {
  __typename?: "Query";
  /** Access to subgraph metadata */
  _meta?: Maybe<_Meta_>;
  auction?: Maybe<Auction>;
  auctions: Array<Auction>;
  batch?: Maybe<Batch>;
  batches: Array<Batch>;
  bid?: Maybe<Bid>;
  bids: Array<Bid>;
  bundle?: Maybe<Bundle>;
  bundles: Array<Bundle>;
  collateralLiquidated?: Maybe<CollateralLiquidated>;
  collateralLiquidateds: Array<CollateralLiquidated>;
  collateralToken?: Maybe<CollateralToken>;
  collateralTokens: Array<CollateralToken>;
  currencyToken?: Maybe<CurrencyToken>;
  currencyTokens: Array<CurrencyToken>;
  deposit?: Maybe<Deposit>;
  deposited?: Maybe<Deposited>;
  depositeds: Array<Deposited>;
  deposits: Array<Deposit>;
  liquidation?: Maybe<Liquidation>;
  liquidations: Array<Liquidation>;
  loan?: Maybe<Loan>;
  loanLiquidated?: Maybe<LoanLiquidated>;
  loanLiquidateds: Array<LoanLiquidated>;
  loanOriginated?: Maybe<LoanOriginated>;
  loanOriginateds: Array<LoanOriginated>;
  loanRepaid?: Maybe<LoanRepaid>;
  loanRepaids: Array<LoanRepaid>;
  loans: Array<Loan>;
  pool?: Maybe<Pool>;
  poolDayData?: Maybe<PoolDayData>;
  poolDayDatas: Array<PoolDayData>;
  poolEvent?: Maybe<PoolEvent>;
  poolEvents: Array<PoolEvent>;
  pools: Array<Pool>;
  redeemed?: Maybe<Redeemed>;
  redeemeds: Array<Redeemed>;
  redemption?: Maybe<Redemption>;
  redemptions: Array<Redemption>;
  tick?: Maybe<Tick>;
  ticks: Array<Tick>;
  tokenCreated?: Maybe<TokenCreated>;
  tokenCreateds: Array<TokenCreated>;
  withdrawn?: Maybe<Withdrawn>;
  withdrawns: Array<Withdrawn>;
};
// ...
export type Redeemed = {
  __typename?: "Redeemed";
  account: Scalars["Bytes"]["output"];
  estimatedAmount: Scalars["BigInt"]["output"];
  id: Scalars["Bytes"]["output"];
  shares: Scalars["BigInt"]["output"];
  tick: Tick;
};
// ...
export type Redemption = {
  __typename?: "Redemption";
  deposit: Deposit;
  id: Scalars["Bytes"]["output"];
  redemptionId: Scalars["BigInt"]["output"];
  shares: Scalars["BigInt"]["output"];
};
// ...
export type Tick = {
  __typename?: "Tick";
  accrualRate?: Maybe<Scalars["BigInt"]["output"]>;
  accrualTimestamp?: Maybe<Scalars["BigInt"]["output"]>;
  accrued?: Maybe<Scalars["BigInt"]["output"]>;
  active: Scalars["Boolean"]["output"];
  available: Scalars["BigInt"]["output"];
  duration: Scalars["BigInt"]["output"];
  durationIndex: Scalars["Int"]["output"];
  id: Scalars["Bytes"]["output"];
  interestWeightedMaturity: Scalars["BigInt"]["output"];
  limit: Scalars["BigInt"]["output"];
  next: Scalars["BigInt"]["output"];
  pending: Scalars["BigInt"]["output"];
  pool: Pool;
  prev: Scalars["BigInt"]["output"];
  principalWeightedDuration: Scalars["BigInt"]["output"];
  rate: Scalars["BigInt"]["output"];
  rateIndex: Scalars["Int"]["output"];
  raw: Scalars["BigInt"]["output"];
  redemptionPending: Scalars["BigInt"]["output"];
  shares: Scalars["BigInt"]["output"];
  token?: Maybe<CurrencyToken>;
  value: Scalars["BigInt"]["output"];
};
// ...
export type TokenCreated = {
  __typename?: "TokenCreated";
  id: Scalars["Bytes"]["output"];
  tick: Tick;
  token: CurrencyToken;
};
// ...
export type Withdrawn = {
  __typename?: "Withdrawn";
  account: Scalars["Bytes"]["output"];
  amount: Scalars["BigInt"]["output"];
  id: Scalars["Bytes"]["output"];
  shares: Scalars["BigInt"]["output"];
  tick: Tick;
};
// ...
export type _Meta_ = {
  __typename?: "_Meta_";
  /**
   * Information about a specific subgraph block. The hash of the block
   * will be null if the _meta field has a block constraint that asks for
   * a block number. It will be filled if the _meta field has no block constraint
   * and therefore asks for the latest  block
   *
   */
  block: _Block_;
  /** The deployment ID */
  deployment: Scalars["String"]["output"];
  /** If `true`, the subgraph encountered indexing errors at some past block */
  hasIndexingErrors: Scalars["Boolean"]["output"];
};

export enum _SubgraphErrorPolicy_ {
  /** Data will be returned even if the subgraph has indexing errors */
  Allow = "allow",
  /** If the subgraph has indexing errors, data will be omitted. The default. */
  Deny = "deny",
}

export type HomePageQueryVariables = Exact<{ [key: string]: never }>;

export type HomePageQuery = {
  __typename?: "Query";
  pool?: {
    __typename?: "Pool";
    id: any;
    totalValueLocked: any;
    collateralToken: { __typename?: "CollateralToken"; id: any; name: string };
    currencyToken: {
      __typename?: "CurrencyToken";
      id: any;
      name: string;
      symbol: string;
    };
    ticks: Array<{
      __typename?: "Tick";
      id: any;
      raw: any;
      duration: any;
      limit: any;
      value: any;
      available: any;
    }>;
  } | null;
};